==> barryma_rapport_stage_1/project_covid19_final/notebooks/07_Evaluation.ipynb <==
import pandas as pd
import logging
import joblib
from sklearn.metrics import mean_squared_error, r2_score

# Setup logging
logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s')

# Load test data and model
logging.info("Loading test data and model")
final_dataset = pd.read_csv("datasets/final_dataset_with_features.csv")
model = joblib.load('models/random_forest_model.pkl')

X_test = final_dataset.drop(columns=['NewCases_7day_avg'])
y_test = final_dataset['NewCases_7day_avg']

# Make predictions
logging.info("Making predictions")
y_pred = model.predict(X_test)

# Evaluate the model
logging.info("Evaluating the model")
mse = mean_squared_error(y_test, y_pred)
r2 = r2_score(y_test, y_pred)

print(f"Mean Squared Error: {mse}")
print(f"R^2 Score: {r2}")
